refactor: Custom exception 추가
